+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- desktop: make selection of multiple dives reasonable fast
 - mobile: add GF fields to adjust Buhlmann algorithm parameters for calculated ceiling
 - undo: save to git after editing weights [#3159]
 - undo: reset dive-mode on undo of set-point addition